+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Renamed affine `matrix` field to `submatrix`.
+
 - Made the `FloatExt` trait sealed.
 
 - Changed the addition/multiplication order of `element_sum` and